+++ uui/src/lib/shared/theme/config/theme-config.types.d.ts
@@ -1,3 +1,10 @@
+import { Properties, PropertiesHyphen } from 'csstype';
+
+export type CssProperties = Properties & PropertiesHyphen;
+
 export interface CssObject {
-  [key: string]: Record<string, string>;
+  [key: string]: CssProperties;
 }
+
+export type BaseScopedStyle = CssProperties | ScopedStyles;
+export type ScopedStyles = { $self?: BaseScopedStyle; [key: string]: BaseScopedStyle };